+++ src/schema/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Clarify format of `PlayerMessage` event when `say` command is used.
+
 ## 0.1.0-beta.2
 
 ### Fixed